Remove old version of TestAccVcdRightsContainers
